refactor: Split ContentWrapper and ContentWrapperResult
To use 2 components instead of one for more clarity
