+++ src/components/ContentWrapper.jsx
@@ -1,50 +1,19 @@
 import React from 'react'
 import { PropTypes } from 'prop-types'
 
-import { Query, isQueryLoading } from 'cozy-client'
-import { Content } from 'cozy-ui/transpiled/react/Layout'
+import { Query } from 'cozy-client'
 
-import Header from './Header'
-import Toolbar from './Toolbar'
-import ContactsList from './ContactsList/ContactsList.jsx'
 import SpinnerContact from './Common/Spinner'
+import ContentResult from './ContentResult'
 import {
   contactsByFamilyNameGivenNameEmailCozyUrl,
   contactsWithoutIndexes
 } from '../helpers/queries'
-import { reworkContacts } from '../helpers/contacts'
-
-export const ContentWrapperResult = ({
-  hasServiceBeenLaunched,
-  contactsWithIndexesResult,
-  contactsWithNoIndexesResult
-}) => {
-  const dataHaveBeenLoaded =
-    !isQueryLoading(contactsWithIndexesResult) &&
-    !isQueryLoading(contactsWithNoIndexesResult) &&
-    !contactsWithIndexesResult.hasMore &&
-    !contactsWithNoIndexesResult.hasMore
-
-  if (!dataHaveBeenLoaded)
-    return <SpinnerContact size="xxlarge" loadingType="fetching_contacts" />
-
-  const contacts = reworkContacts(
-    hasServiceBeenLaunched,
-    contactsWithIndexesResult.data,
-    contactsWithNoIndexesResult.data
-  )
-
-  return (
-    <>
-      {contacts.length >= 1 && <Header right={<Toolbar />} />}
-      <Content>
-        <ContactsList contacts={contacts} />
-      </Content>
-    </>
-  )
-}
 
 const ContentWrapper = ({ hasServiceBeenLaunched }) => {
+  if (hasServiceBeenLaunched === null) {
+    return <SpinnerContact size="xxlarge" loadingType="fetching_contacts" />
+  }
   return (
     <Query
       query={contactsByFamilyNameGivenNameEmailCozyUrl.definition}
@@ -66,7 +35,7 @@ const ContentWrapper = ({ hasServiceBeenLaunched }) => {
               }
 
               return (
-                <ContentWrapperResult
+                <ContentResult
                   hasServiceBeenLaunched={hasServiceBeenLaunched}
                   contactsWithIndexesResult={contactsWithIndexesResult}
                   contactsWithNoIndexesResult={contactsWithNoIndexesResult}
